requirements.txt
requests
notion-client
retrying
openai
python-dotenv